feat: provide allocator
While semver is no_std, it still requires an allocator. Furthermore wasm32v1-none doesn't support std. And the initial implementation of alloc in the soroban_sdk required the use of `extern crate std`. This was fixed in the newest version of the soroban-sdk but we need to update the loam-sdk to handle this change. So for now can ship the same allocator implementation soroban-sdk provides. Once updated we can remove it.