fix: build clients race condition
Previously, the client build process deleted the old contract code
before building and installing the new ones. While running in watch mode
concurrently with the frontend dev server, this caused a hot-reload
which would throw an error since app code is importing contracts that
don't exist until rebuilt.

This change builds and installs the contracts in a temp directory then
replaces the contract code at once, ensuring only one hot-reload and no
broken imports. This has the added benefit of being able to use the old
contract code as a backup if the build/install process fails for any
reason.