Fix: Build clients race condition (#126)
* fix: build clients race condition

Previously, the client build process deleted the old contract code
before building and installing the new ones. While running in watch mode
concurrently with the frontend dev server, this caused a hot-reload
which would throw an error since app code is importing contracts that
don't exist until rebuilt.

This change builds and installs the contracts in a temp directory then
replaces the contract code at once, ensuring only one hot-reload and no
broken imports. This has the added benefit of being able to use the old
contract code as a backup if the build/install process fails for any
reason.

* fix: lint error

* fix: remove redundant temp copy

* fix: use overwrite

* fix: create initial packages folder if it doesn't exist

* fix: only copy over the artifacts when updating the bindings

* make sure to write contract client  helper code

* update tests

---------

Co-authored-by: Willem Wyndham <willem@ahalabs.dev>
Co-authored-by: Blaine Heffron <baheffron@gmail.com>