fix(build-clients): only `allowHttp` in dev & test
This got switched around by mistake in 1bb8c7e35af9b5632f0515aab5c5c929fe58a870

We need tests on this! But for now it's a show-stopping bug (no local
environment works correctly without this), and I have verified that it
works correctly by using it in a local project.